test(smoke): keep the Bun drivers off browser-only imports
media.ts pulled FAULTS out of fixture.ts, which imports @moq/publish, whose audio worklet is a vite-only asset: bun failed to load it before any check ran. Move everything both sides need into src/contract.ts, which imports nothing from the browser packages.
